requirements-cuda12.txt
numpy
tqdm
h5py
cupy-cuda12x